requirements.txt
# --- Core Framework & Orchestration ---
langchain==1.2.0
langchain-core==1.2.5
langchain-community==0.4.1
langchain-experimental==0.4.1
langchain-classic==1.0.1
langchain-chroma==1.1.0  # Vector Store
langchain-google-genai==4.1.2  # LLM (Gemini)
langchain-huggingface==1.2.0  # Embeddings

# --- Vector Database & Search ---
chromadb==1.4.0
rank-bm25==0.2.2  # Hybrid Search (BM25)
FlashRank==0.2.10  # Reranking

# --- Machine Learning & Embeddings ---
torch==2.5.1+cu121  # Check CPU/CUDA version compatibility
sentence-transformers==5.2.0

# --- Web UI ---
streamlit==1.52.2

# --- Data Processing & Validation ---
pydantic==2.12.5
pydantic-settings==2.12.0
sqlalchemy==2.0.45  # Ingestion state management

# --- Document Parsing & Crawling ---
beautifulsoup4==4.14.3
lxml==6.0.2
pdfplumber==0.11.8
markdownify==1.2.2
requests==2.32.5
selenium==4.39.0
webdriver-manager==4.0.2
aiofiles==25.1.0  # Async file I/O

# --- Utilities & Logging ---
python-dotenv==1.2.1
loguru==0.7.3
rich==14.2.0
tenacity==9.1.2
tqdm==4.67.1
